restructure calibration for more modularity
